let's try to compile this on a windows machine

+++ src/host/asio/stream.rs
@@ -298,7 +298,7 @@ impl Device {
                         callback_info,
                         config.sample_rate,
                         std::convert::identity::<f64>,
-                        std::convert::identity::<f64>,
+                        std::convert::identity::<f32>,
                     );
                 }
